requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-countries==5.5
django-crispy-forms==1.8.1
django-dotenv==1.4.2
django-formtools==2.2
django-getenv==1.3.2
django-language-field==0.0.3
django-tempus-dominus==5.1.2.13
djangorestframework==3.11.0
fake-factory==9999.9.9
Faker==4.0.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
stripe==2.42.0
text-unidecode==1.3
urllib3==1.25.7
wrapt==1.11.2
